Add Order Endpoints and API to adrress

+++ src/products/products.controller.ts
@@ -1,6 +1,9 @@
-import { Controller, Get, Param, Delete, Post, Body } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Post, Body, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
-import { db, Product } from './../db';
+import { ParseUUIDPipe, NotFoundException } from '@nestjs/common';
+import { CreateProductDTO } from './dtos/create.product.dto';
+import { UpdateProductDTO } from "./dtos/update-product.dto";
+
 @Controller('products')
 export class ProductsController {
   constructor(private productsService: ProductsService) {}
@@ -9,16 +12,32 @@ export class ProductsController {
     return this.productsService.getAll();
   }
   @Get('/:id')
-  public getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
+
   @Delete('/:id')
-  deleteById(@Param('id') id: string) {
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteById(id);
     return { success: true };
   }
   @Post('/')
-  create(@Body() productData: Omit<Product, 'id'>) {
+  create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
   }
 }
